设计模式之Factory系列，FactoryMethod, AbstractFactory，以及spring bean工厂开头
