--- src/com/mashibing/singleton/Main.java
@@ -1,8 +1,0 @@
-package com.mashibing.singleton;
-
-public class Main {
-
-    public static void main(String[] args) {
-        //Mgr01 mgr = Mgr01.getInstance();
-    }
-}
