--- src/com/mashibing/strategy/CatHeightComparator.java
@@ -1,10 +1,0 @@
-package com.mashibing.strategy;
-
-public class CatHeightComparator implements Comparator<Cat> {
-    @Override
-    public int compare(Cat o1, Cat o2) {
-        if(o1.height > o2.height) return -1;
-        else if (o1.height < o2.height) return 1;
-        else return 0;
-    }
-}
