--- src/com/mashibing/strategy/Comparable.java
@@ -1,5 +1,0 @@
-package com.mashibing.strategy;
-
-public interface Comparable<T> {
-    int compareTo(T o);
-}
