--- src/com/mashibing/strategy/Comparator.java
@@ -1,5 +1,0 @@
-package com.mashibing.strategy;
-
-public interface Comparator<T> {
-    int compare(T o1, T o2);
-}
