--- src/com/mashibing/strategy/Dog.java
@@ -1,24 +1,0 @@
-package com.mashibing.strategy;
-
-public class Dog implements Comparable<Dog> {
-
-    int food;
-
-    public Dog(int food) {
-        this.food = food;
-    }
-
-    @Override
-    public int compareTo(Dog d) {
-        if(this.food < d.food) return -1;
-        else if(this.food > d.food) return 1;
-        else return 0;
-    }
-
-    @Override
-    public String toString() {
-        return "Dog{" +
-                "food=" + food +
-                '}';
-    }
-}
