+++ src/main/java/com/mashibing/dp/abstractfactory/AK47.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.abstractfactory;
+
+public class AK47 extends Weapon{
+    public void shoot() {
+        System.out.println("tututututu....");
+    }
+}
