+++ src/main/java/com/mashibing/dp/abstractfactory/AbastractFactory.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.abstractfactory;
+
+public abstract class AbastractFactory {
+    abstract Food createFood();
+    abstract Vehicle createVehicle();
+    abstract Weapon createWeapon();
+}
