+++ src/main/java/com/mashibing/dp/abstractfactory/Bread.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.abstractfactory;
+
+public class Bread extends Food{
+    public void printName() {
+        System.out.println("wdm");
+    }
+}
