+++ src/main/java/com/mashibing/dp/abstractfactory/Broom.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.abstractfactory;
+
+public class Broom extends Vehicle{
+    public void go() {
+        System.out.println("Car go wuwuwuwuw....");
+    }
+}
