+++ src/main/java/com/mashibing/dp/abstractfactory/Car.java
@@ -1,0 +1,8 @@
+package com.mashibing.dp.abstractfactory;
+
+public class Car extends Vehicle{
+
+    public void go() {
+        System.out.println("Car go wuwuwuwuw....");
+    }
+}
