+++ src/main/java/com/mashibing/dp/abstractfactory/Food.java
@@ -1,0 +1,5 @@
+package com.mashibing.dp.abstractfactory;
+
+public abstract class Food {
+   abstract void printName();
+}
