+++ src/main/java/com/mashibing/dp/abstractfactory/MagicFactory.java
@@ -1,0 +1,18 @@
+package com.mashibing.dp.abstractfactory;
+
+public class MagicFactory extends AbastractFactory {
+    @Override
+    Food createFood() {
+        return new MushRoom();
+    }
+
+    @Override
+    Vehicle createVehicle() {
+        return new Broom();
+    }
+
+    @Override
+    Weapon createWeapon() {
+        return new MagicStick();
+    }
+}
