+++ src/main/java/com/mashibing/dp/abstractfactory/MagicStick.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.abstractfactory;
+
+public class MagicStick extends Weapon{
+    public void shoot() {
+        System.out.println("diandian....");
+    }
+}
