+++ src/main/java/com/mashibing/dp/abstractfactory/ModernFactory.java
@@ -1,0 +1,19 @@
+package com.mashibing.dp.abstractfactory;
+
+
+public class ModernFactory extends AbastractFactory {
+    @Override
+    Food createFood() {
+        return new Bread();
+    }
+
+    @Override
+    Vehicle createVehicle() {
+        return new Car();
+    }
+
+    @Override
+    Weapon createWeapon() {
+        return new AK47();
+    }
+}
