+++ src/main/java/com/mashibing/dp/abstractfactory/MushRoom.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.abstractfactory;
+
+public class MushRoom extends Food{
+    public void printName() {
+        System.out.println("dmg");
+    }
+}
