+++ src/main/java/com/mashibing/dp/abstractfactory/Vehicle.java
@@ -1,0 +1,5 @@
+package com.mashibing.dp.abstractfactory;
+
+public abstract class Vehicle { //interface
+    abstract void go();
+}
