+++ src/main/java/com/mashibing/dp/abstractfactory/Weapon.java
@@ -1,0 +1,5 @@
+package com.mashibing.dp.abstractfactory;
+
+public abstract class Weapon {
+    abstract void shoot();
+}
