+++ src/main/java/com/mashibing/dp/cor/servlet/Servlet_Main.java
@@ -1,0 +1,79 @@
+package com.mashibing.dp.cor.servlet;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Servlet_Main {
+    public static void main(String[] args) {
+        Request request = new Request();
+        request.str = "request";
+        Response response = new Response();
+        response.str = "response";
+
+        System.out.println();
+
+    }
+}
+
+class Msg {
+    private String name;
+    private String msg;
+
+    public String getMsg() {
+        return msg;
+    }
+
+    public void setMsg(String msg) {
+        this.msg = msg;
+    }
+
+    @Override
+    public String toString() {
+        return "Msg{" +
+                "msg='" + msg + '\'' +
+                '}';
+    }
+}
+
+interface Filter {
+    boolean doFilter(Request equest,Response response);
+}
+
+class HTMLFilter implements Filter {
+    @Override
+    public boolean doFilter(Request equest,Response response) {
+        //处理request
+        //处理response的str
+        return false;
+    }
+}
+
+class Request {
+    String str;
+}
+
+class Response {
+    String str;
+}
+
+class SensitiveFilter implements Filter {
+    @Override
+    public boolean doFilter(Request equest,Response response) {
+        return false;
+    }
+}
+
+
+
+class FilterChain implements Filter {
+    List<Filter> filters = new ArrayList<>();
+
+    public FilterChain add(Filter f) {
+        filters.add(f);
+        return this;
+    }
+
+    public boolean doFilter(Request equest,Response response) {
+        return false;
+    }
+}
