+++ src/main/java/com/mashibing/dp/factorymethod/Broom.java
@@ -1,0 +1,9 @@
+package com.mashibing.dp.factorymethod;
+
+public class Broom implements Moveable{
+
+    @Override
+    public void go() {
+        System.out.println("broom flying chuachuachua .....");
+    }
+}
