+++ src/main/java/com/mashibing/dp/factorymethod/Car.java
@@ -1,0 +1,8 @@
+package com.mashibing.dp.factorymethod;
+
+public class Car implements  Moveable {
+
+    public void go() {
+        System.out.println("Car go wuwuwuwuw....");
+    }
+}
