+++ src/main/java/com/mashibing/dp/factorymethod/CarFactory.java
@@ -1,0 +1,8 @@
+package com.mashibing.dp.factorymethod;
+
+public class CarFactory {
+    public Moveable create() {
+        System.out.println("a car created!");
+        return new Car();
+    }
+}
