+++ src/main/java/com/mashibing/dp/factorymethod/Main.java
@@ -1,0 +1,8 @@
+package com.mashibing.dp.factorymethod;
+
+public class Main {
+    public static void main(String[] args) {
+        Moveable m = new CarFactory().createCar();
+        m.go();
+    }
+}
