+++ src/main/java/com/mashibing/dp/factorymethod/Moveable.java
@@ -1,0 +1,5 @@
+package com.mashibing.dp.factorymethod;
+
+public interface Moveable {
+    void go();
+}
