+++ src/main/java/com/mashibing/dp/factorymethod/Plane.java
@@ -1,0 +1,7 @@
+package com.mashibing.dp.factorymethod;
+
+public class Plane implements Moveable {
+    public void go() {
+        System.out.println("plane flying shushua....");
+    }
+}
