+++ src/main/java/com/mashibing/dp/factorymethod/SimpleVehicleFactory.java
@@ -1,0 +1,15 @@
+package com.mashibing.dp.factorymethod;
+
+/**
+ * 简单工厂的可扩展性不好
+ */
+public class SimpleVehicleFactory {
+    public Car createCar() {
+        //before processing
+        return new Car();
+    }
+
+    public Broom createBroom() {
+        return new Broom();
+    }
+}
