+++ src/main/java/com/mashibing/dp/singleton/Main.java
@@ -1,0 +1,8 @@
+package com.mashibing.dp.singleton;
+
+public class Main {
+
+    public static void main(String[] args) {
+        //Mgr01 mgr = Mgr01.getInstance();
+    }
+}
