+++ src/main/java/com/mashibing/dp/strategy/CatWeightComparator.java
@@ -1,0 +1,10 @@
+package com.mashibing.dp.strategy;
+
+public class CatWeightComparator implements Comparator<Cat> {
+    @Override
+    public int compare(Cat o1, Cat o2) {
+        if(o1.weight < o2.weight) return -1;
+        else if (o1.weight > o2.weight) return 1;
+        else return 0;
+    }
+}
