+++ src/main/java/com/mashibing/dp/strategy/Comparable.java
@@ -1,0 +1,5 @@
+package com.mashibing.dp.strategy;
+
+public interface Comparable<T> {
+    int compareTo(T o);
+}
