+++ src/main/java/com/mashibing/dp/strategy/Comparator.java
@@ -1,0 +1,10 @@
+package com.mashibing.dp.strategy;
+
+@FunctionalInterface
+public interface Comparator<T> {
+    int compare(T o1, T o2);
+
+    default void m() {
+        System.out.println("m");
+    }
+}
